fix(fe): move files location and change arrow direction
