+++ apps/frontend/app/admin/course/[courseId]/assignment/[assignmentId]/(overall)/@tabs/submission/page.tsx
@@ -3,9 +3,8 @@
 import { FetchErrorFallback } from '@/components/FetchErrorFallback'
 import { cn } from '@/libs/utils'
 import { ErrorBoundary } from '@suspensive/react'
-import { useState } from 'react'
-import { Suspense } from 'react'
-import { ParticipantTable } from '../_components/ParticipantTable'
+import { Suspense, useState } from 'react'
+import { ParticipantTable } from './_components/ParticipantTable'
 import {
   SubmissionTable,
   SubmissionTableFallback
